+++ sdk/openai/openai/test/public/openai.spec.ts
@@ -7,19 +7,19 @@ import { Context } from "mocha";
 import { AuthMethod, createClient, startRecorder } from "./utils/recordedClient.js";
 import {
   assertChatCompletions,
-  assertChatCompletionsStream,
+  assertChatCompletionsList,
   assertCompletions,
   assertCompletionsStream,
 } from "./utils/asserts.js";
 import {
+  bufferAsyncIterable,
   getDeployments,
   getModels,
   getSucceeded,
   sendRequestWithRecorder,
   updateWithSucceeded,
-  withDeployment,
+  withDeployments,
 } from "./utils/utils.js";
-import { logger } from "./utils/logger.js";
 import { createHttpHeaders, createPipelineRequest } from "@azure/core-rest-pipeline";
 import { getImageDimensions } from "./utils/getImageDimensions.js";
 import { OpenAIClient, ImageLocation } from "../../src/index.js";
@@ -54,10 +54,10 @@ describe("OpenAI", function () {
       describe("getCompletions", function () {
         it("returns completions across all models", async function () {
           const prompt = ["What is Azure OpenAI?"];
-          await withDeployment(
+          await withDeployments(
             authMethod === "OpenAIKey" ? models : deployments,
-            async (deploymentName) =>
-              client.getCompletions(deploymentName, prompt).then(assertCompletions)
+            async (deploymentName) => client.getCompletions(deploymentName, prompt),
+            assertCompletions
           );
         });
       });
@@ -183,7 +183,7 @@ describe("OpenAI", function () {
         describe("getChatCompletions", function () {
           it("returns completions across all models", async function () {
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -191,10 +191,8 @@ describe("OpenAI", function () {
                   chatCompletionDeployments,
                   chatCompletionModels
                 ),
-                async (deploymentName) =>
-                  client
-                    .getChatCompletions(deploymentName, pirateMessages)
-                    .then(assertChatCompletions)
+                async (deploymentName) => client.getChatCompletions(deploymentName, pirateMessages),
+                assertChatCompletions
               ),
               chatCompletionDeployments,
               chatCompletionModels,
@@ -204,7 +202,7 @@ describe("OpenAI", function () {
 
           it("calls functions", async function () {
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -213,11 +211,10 @@ describe("OpenAI", function () {
                   chatCompletionModels
                 ),
                 async (deploymentName) =>
-                  client
-                    .getChatCompletions(deploymentName, weatherMessages, {
-                      functions: [getCurrentWeather],
-                    })
-                    .then((c) => assertChatCompletions(c, { functions: true }))
+                  client.getChatCompletions(deploymentName, weatherMessages, {
+                    functions: [getCurrentWeather],
+                  }),
+                (c) => assertChatCompletions(c, { functions: true })
               ),
               chatCompletionDeployments,
               chatCompletionModels,
@@ -225,12 +222,12 @@ describe("OpenAI", function () {
             );
           });
 
-          it("bring your own data", async function (this: Context) {
+          it.skip("bring your own data", async function (this: Context) {
             if (authMethod === "OpenAIKey") {
               this.skip();
             }
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -239,22 +236,21 @@ describe("OpenAI", function () {
                   chatCompletionModels
                 ),
                 async (deploymentName) =>
-                  client
-                    .getChatCompletions(deploymentName, byodMessages, {
-                      azureExtensionOptions: {
-                        extensions: [
-                          {
-                            type: "AzureCognitiveSearch",
-                            parameters: {
-                              endpoint: assertEnvironmentVariable("AZURE_SEARCH_ENDPOINT"),
-                              key: assertEnvironmentVariable("AZURE_SEARCH_KEY"),
-                              indexName: assertEnvironmentVariable("AZURE_SEARCH_INDEX"),
-                            },
+                  client.getChatCompletions(deploymentName, byodMessages, {
+                    azureExtensionOptions: {
+                      extensions: [
+                        {
+                          type: "AzureCognitiveSearch",
+                          parameters: {
+                            endpoint: assertEnvironmentVariable("AZURE_SEARCH_ENDPOINT"),
+                            key: assertEnvironmentVariable("AZURE_SEARCH_KEY"),
+                            indexName: assertEnvironmentVariable("AZURE_SEARCH_INDEX"),
                           },
-                        ],
-                      },
-                    })
-                    .then(assertChatCompletions)
+                        },
+                      ],
+                    },
+                  }),
+                assertChatCompletions
               ),
               chatCompletionDeployments,
               chatCompletionModels,
@@ -266,7 +262,7 @@ describe("OpenAI", function () {
         describe("listChatCompletions", function () {
           it("returns completions across all models", async function () {
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -274,19 +270,17 @@ describe("OpenAI", function () {
                   chatCompletionDeployments,
                   chatCompletionModels
                 ),
-                async (deploymentName) => {
-                  const count = await assertChatCompletionsStream(
-                    client.listChatCompletions(deploymentName, pirateMessages),
-                    {
-                      // The API returns an empty choice in the first event for some
-                      // reason. This should be fixed in the API.
-                      allowEmptyChoices: true,
-                    }
-                  );
-                  if (count === 0) {
-                    logger.warning(`No completions returned for ${deploymentName}`);
-                  }
-                }
+                async (deploymentName) =>
+                  bufferAsyncIterable(client.listChatCompletions(deploymentName, pirateMessages)),
+                async (res) =>
+                  assertChatCompletionsList(res, {
+                    // The API returns an empty choice in the first event for some
+                    // reason. This should be fixed in the API.
+                    allowEmptyChoices: true,
+                    // The API returns an empty ID in the first event for some
+                    // reason. This should be fixed in the API.
+                    allowEmptyId: true,
+                  })
               ),
               chatCompletionDeployments,
               chatCompletionModels,
@@ -296,7 +290,7 @@ describe("OpenAI", function () {
 
           it("calls functions", async function () {
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -304,22 +298,19 @@ describe("OpenAI", function () {
                   chatCompletionDeployments,
                   chatCompletionModels
                 ),
-                async (deploymentName) => {
-                  const count = await assertChatCompletionsStream(
+                async (deploymentName) =>
+                  bufferAsyncIterable(
                     client.listChatCompletions(deploymentName, weatherMessages, {
                       functions: [getCurrentWeather],
-                    }),
-                    {
-                      functions: true,
-                      // The API returns an empty choice in the first event for some
-                      // reason. This should be fixed in the API.
-                      allowEmptyChoices: true,
-                    }
-                  );
-                  if (count === 0) {
-                    logger.warning(`No completions returned for ${deploymentName}`);
-                  }
-                }
+                    })
+                  ),
+                (res) =>
+                  assertChatCompletionsList(res, {
+                    functions: true,
+                    // The API returns an empty choice in the first event for some
+                    // reason. This should be fixed in the API.
+                    allowEmptyChoices: true,
+                  })
               ),
               chatCompletionDeployments,
               chatCompletionModels,
@@ -327,12 +318,12 @@ describe("OpenAI", function () {
             );
           });
 
-          it("bring your own data", async function () {
+          it.skip("bring your own data", async function () {
             if (authMethod === "OpenAIKey") {
               this.skip();
             }
             updateWithSucceeded(
-              await withDeployment(
+              await withDeployments(
                 getSucceeded(
                   authMethod,
                   deployments,
@@ -340,8 +331,8 @@ describe("OpenAI", function () {
                   chatCompletionDeployments,
                   chatCompletionModels
                 ),
-                async (deploymentName) => {
-                  const count = await assertChatCompletionsStream(
+                async (deploymentName) =>
+                  bufferAsyncIterable(
                     client.listChatCompletions(deploymentName, byodMessages, {
                       azureExtensionOptions: {
                         extensions: [
@@ -356,11 +347,8 @@ describe("OpenAI", function () {
                         ],
                       },
                     })
-                  );
-                  if (count === 0) {
-                    logger.warning(`No completions returned for ${deploymentName}`);
-                  }
-                }
+                  ),
+                assertChatCompletionsList
               ),
               chatCompletionDeployments,
               chatCompletionModels,